requirements.txt
asgiref==3.4.1
autopep8==1.6.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cssselect==1.1.0
Django==4.0
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.4.0
gunicorn==20.1.0
idna==3.3
jieba3k==0.35.1
joblib==1.1.0
lxml==4.7.1
newspaper3k==0.2.8
nltk==3.6.6
Pillow==8.4.0
pycodestyle==2.8.0
python-dateutil==2.8.2
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.3.1
spotipy==2.19.0
sqlparse==0.4.2
tinysegmenter==0.3
tldextract==3.1.2
toml==0.10.2
tqdm==4.62.3
urllib3==1.26.7
